\docType{package}
\name{faoutlier-package}
\alias{faoutlier-package}
\title{Influential case detection methods for FA and SEM}
\description{
  Influential case detection methods for factor analysis
  and SEM

  Influential case detection methods for factor analysis
  and SEM
}
\details{
  Implements robust Mahalanobis methods, generalized Cook's
  distances, likelihood ratio tests, model implied
  residuals, and variaous graphical methods to help detect
  abd summarize oinfluential cases that can affect factor
  analysis and structural equation modeling.

  Implements robust Mahalanobis methods, generalized Cook's
  distances, likelihood ratio tests, model implied
  residuals, and variaous graphical methods to help detect
  abd summarize oinfluential cases that can affect factor
  analysis and structural equation modeling.
}
\author{
  \email{rphilip.chalmers@gmail.com}

  \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

