% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoutlier.R
\name{faoutlier}
\alias{faoutlier}
\alias{faoutlier-package}
\title{Influential case detection methods for FA and SEM}
\description{
Influential case detection methods for factor analysis and SEM
}
\details{
Implements robust Mahalanobis methods, generalized Cook's distances,
likelihood ratio tests, model implied residuals, and various
graphical methods to help detect and summarize influential
cases that can affect exploratory and confirmatory factor analyses.
}
\references{
Chalmers, R. P. & Flora, D. B. (2015). faoutlier: An R Package for Detecting
  Influential Cases in Exploratory and Confirmatory Factor Analysis.
  \emph{Applied Psychological Measurement, 39}, 573-574. \doi{10.1177/0146621615597894}

Flora, D. B., LaBrish, C. & Chalmers, R. P. (2012). Old and new ideas for data
screening and assumption testing for exploratory and confirmatory factor analysis.
 \emph{Frontiers in Psychology, 3}, 1-21. \doi{10.3389/fpsyg.2012.00055}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}
