% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{faoutlier}
\alias{faoutlier}
\alias{faoutlier-package}
\title{Influential case detection methods for FA and SEM}
\description{
Influential case detection methods for factor analysis and SEM
}
\details{
Implements robust Mahalanobis methods, generalized Cook's distances,
likelihood ratio tests, model implied residuals, and various
graphical methods to help detect and summarize influential
cases that can affect exploratory and confirmatory factor analyses.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

