\docType{data}
\name{holzinger.outlier}
\alias{holzinger.outlier}
\title{Description of holzinger data with 1 outlier}
\description{
  A sample of 100 simulated cases from the infamous
  Holzinger dataset using 9 variables, but with 1 outlier
  added to the dataset. The first row was replaced by
  adding 2 to five of the observed variables (odd-numbered
  items) and subtracting 2 from the other four observed
  variables (even-numbered items).
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
\keyword{datasets}

